requirements.txt
passlib==1.7.2
fastapi==0.54.2
pydantic==1.5.1
tortoise-orm==0.16.11
